# Go dependency download and analysis performance fix